+++ src/TaskWorker/TaskWorker.cpp
@@ -2,7 +2,7 @@
 
 #include "Enums.hpp"
 #include "Utils.hpp"
-#include "rpgmtranslate.h"
+#include "rpgmtranslate.hpp"
 
 #include <QApplication>
 #include <QFile>